+++ src/app/page.scss
@@ -426,11 +426,13 @@
 
 .alert {
   background-color: transparent;
-  border: 1px solid $beige;
-  color: $beige;
+  border: 1px solid $gold;
+  border-radius: 3px;
+  color: $fontColor;
   font-size: 1rem;
   margin-bottom: 2rem;
   opacity: 0.8;
+  padding: 16px;
   width: 100%;
 }
 .form {
