+++ src/app/page.tsx
@@ -3,6 +3,7 @@ import React, { useEffect, useState } from "react";
 import { useRouter } from"next/navigation";
 import Image from "next/image";
 import Link from "next/link";
+import Form from "@/components/Form/Form";
 import { BsFacebook } from "react-icons/bs";
 import { BsInstagram } from "react-icons/bs";
 import { BsFillTelephoneFill } from "react-icons/bs";
@@ -17,7 +18,7 @@ import offer5 from "../assets/offer5.jpg";
 import offer4 from "../assets/offer4.jpg";
 import logo from "../assets/logoDuze.png";
 import "./page.scss";
-import emailjs from "@emailjs/browser";
+
 import LoadingButton from "@/components/LoadingButton/LoadingButton";
 
 const data = [
@@ -49,7 +50,7 @@ const data = [
 
 export default function Home() {
   const [loading, setLoading] = useState<boolean>(false);
-  const [status, setStatus] = useState<boolean>(false);
+
   const [active, setActive] = useState<boolean>(false);
 
   const [activeLink, setActiveLink]=useState<string>('section1');
@@ -82,36 +83,12 @@ export default function Home() {
     router.push(link)
   }
 
-  const sendEmail = async (e: any) => {
-    e.preventDefault();
-    setLoading(true);
-
-    emailjs
-      .sendForm(
-        "service_cegqeco",
-        "template_c9qk8w7",
-        e.target,
-        "2GSQoPpt-ZCUvyQgM"
-      )
-      .then(
-        (result) => {
-          console.log(result.text);
-        },
-        (error) => {
-          console.log(error.text);
-        }
-      );
-
-    setTimeout(() => {
-      setStatus(true);
-      e.target.reset();
-      setLoading(false);
-    }, 200);
-  };
+  
 
   const showMenu = () => {
     setActive(!active);
   };
+
 
   return (
     <>
@@ -263,68 +240,7 @@ export default function Home() {
             </ul>
           </article>
         </section>
-        <section className="contactForm" id="section4">
-          <div className="formArea">
-            <form className="form" onSubmit={sendEmail}>
-              <h5 className="contactForm__heading">Napisz do mnie</h5>
-              <p className="contactForm__content">
-                Zapraszam do kontaktu
-              </p>
-
-              {status ? (
-                <p className="alert">Twoja wiadomość została wysłana</p>
-              ) : null}
-
-              <label htmlFor="name"></label>
-              <input
-                type="text"
-                className="form__control"
-                id="name"
-                name="name"
-                aria-describedby="emailHelp"
-                placeholder="Imię i nazwisko *"
-                required
-              />
-
-              <label htmlFor="email"></label>
-              <input
-                type="email"
-                className="form__control"
-                id="email"
-                name="email"
-                placeholder="Email *"
-                required
-              />
-
-              <label htmlFor="phone"></label>
-              <input
-                type="tel"
-                className="form__control"
-                id="phone"
-                name="phone"
-                placeholder="Telefon *"
-                required
-              />
-
-              <textarea
-                className="form__control"
-                id="message"
-                name="message"
-                placeholder="Twoja wiadomość *"
-                maxLength={200}
-                required
-              />
-              <p className="form__info">* pole wymagane</p>
-              <LoadingButton className="button" loading={loading}>Wyślij</LoadingButton>
-            </form>
-
-            <Image
-              className="formArea__img"
-              src={baby}
-              alt="przykład ciast"
-            />
-          </div>
-        </section>
+        <Form />
       </main>
 
       <footer className="footer">
